Customer Module Second Commit
This module have Customer entity, service, repository

+++ src/main/java/com/onlinesweetmart/service/CustomerService.java
@@ -6,15 +6,16 @@ import com.onlinesweetmart.entity.Customer;
 
 public interface CustomerService {
 	
-	public Customer addCustomer(Customer customer);
+    public Customer addCustomer(Customer customer);
 	
 	public Customer updateCustomer(Customer customer);
 	
-	public void cancelCustomer(int customerId);
+	public void cancelCustomer(long customerId);
 	
 	public List<Customer> showAllCustomers();
 
-	public Customer showAllcustomers(int customerId);
+	public Customer showAllcustomers(long customerId);
+
